Raises error if kit.subkits and kits.features both used
Eons ago, we renamed the concept known as `subkits` to `features`, as we
felt it better reflected the behaviour.  However, we continued to allow
the use of the `subkits` key under the top-level `kit` key.  This has
caused the rare conflict when a hierarchal yml file (such as corp.yaml)
contained `kit.subkits`, and the dependent environment file (such as
corp-myenv.yml) contained `kit.features` (or vice versa).  In this case,
`features` presence would trump `subkits` and the latter would be
ignored.

This commit causes an error to be raised when any mixed usage is
encountered.