Extend -C option to allow specifying yaml file
The -C option allows users to specify an alternative directory as the
base genesis directory instead of the current directory.  As a convenience,
this change allows users to specify the actual environment YAML file
instead of the base directory.  The base directory will then be whatever
ancestor directory above that file that contains the .genesis directory.